usecase_classification/requirements.txt
google-cloud-aiplatform==1.15.0
google-cloud-bigquery==2.34.4
google-cloud-bigquery-storage==2.13.2
google-cloud-storage==1.44.0
kfp==1.8.13
google-cloud-pipeline-components==1.0.14
icecream==2.1.1
pandas-gbq==0.15.0
pandas==1.3.3
google-cloud-pubsub==2.13.4